athena-locust/requirements.txt
locust>=2.39.0
requests
influxdb-client
factory_boy
faker